tests/wpt/tests/navigation-api/commit-behavior/after-transition-new-navigation-before-commit.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../navigation-methods/return-value/resources/helpers.js"></script>
<body>
<script>
promise_test(async t => {
  navigation.addEventListener("navigate", e => e.intercept({ commit: "after-transition" }), { once: "true" });

  let navigateerror_called = false;
  navigation.onnavigateerror = t.step_func(() => {
    navigateerror_called = true;
    assert_equals(location.hash, "");
  });

  let promises_should_not_commit = navigation.navigate("#ShouldNotCommit");
  let promises_fulfilled = navigation.navigate("#1");
  await assertBothRejectDOM(t, promises_should_not_commit, "AbortError");
  await assertBothFulfill(t, promises_fulfilled, navigation.currentEntry);

  assert_equals(location.hash, "#1");
  assert_true(navigateerror_called);
}, "Cancel a { commit: 'after-transition' } navigation before commit() by starting a new navigation");
</script>
</body>
